+++ .flake8
@@ -1,5 +1,5 @@
 [flake8]
-max-line-length = 130
+max-line-length = 145
 exclude =
     .venv,
     __pycache__,
