+++ cli/data_masking/data_masking_rows_generator.py
@@ -2,7 +2,7 @@ from typing import List, Dict
 
 from pydantic import BaseModel
 
-from cli.data_masking.data_masking_config_file_parser import (
+from cli.data_masking.data_masking_config_parser import (
     DataMaskingConfig,
 )
 from cli.model import ManifestNode
